+++ docs/changes.rst
@@ -28,6 +28,7 @@ Not yet released.
 * Allow to configure ``Content-Security-Policy`` HTTP headers.
 * Added support for aliasing languages at project level.
 * New addon to help with HTML or JavaScript localization, see :ref:`addon-weblate.cdn.cdnjs`.
+* The Weblate domain is now configured in the settings, see :setting:`SITE_DOMAIN`.
 
 Weblate 4.1.1
 -------------